move criterion and postprocess to independent scripts
