+++ defm_detr/models/deformable_detr.py
@@ -14,22 +14,17 @@ import torch
 import torch.nn.functional as F
 from torch import nn
 import math
-from typing import List
 
 import settings
-from util import box_ops
-from util.misc import (NestedTensor, nested_tensor_from_tensor_list,
-                       accuracy, get_world_size, interpolate,
-                       is_dist_avail_and_initialized, inverse_sigmoid)
+from util.misc import NestedTensor, nested_tensor_from_tensor_list, inverse_sigmoid
 
 from .timm_models import build_hf_backbone
-from .backbone import build_backbone
+from .criterion import SetCriterion
 from .matcher import build_matcher
-from .segmentation import (DETRsegm, PostProcessPanoptic, PostProcessSegm,
-                           dice_loss, sigmoid_focal_loss)
+from .postprocess import PostProcess
+from .segmentation import DETRsegm, PostProcessSegm
 from .deformable_transformer import build_deforamble_transformer
 import copy
-from config.config import load_config
 from util.misc import MLP
 
 device = torch.device('cuda') if torch.cuda.is_available() else torch.device('cpu')
@@ -200,232 +195,6 @@ class DeformableDETR(nn.Module):
         # as a dict having both a Tensor and a list.
         return [{'pred_logits': a, 'pred_boxes': b}
                 for a, b in zip(outputs_class[:-1], outputs_coord[:-1])]
-
-
-class SetCriterion(nn.Module):
-    ''' This class computes the loss for DETR.
-    The process happens in two steps:
-        1) we compute hungarian assignment between ground truth boxes and the outputs of the model
-        2) we supervise each pair of matched ground-truth / prediction (supervise class and box)
-    '''
-    def __init__(self, num_classes, matcher, loss_names: List[str], focal_alpha=0.25):
-        ''' Create the criterion.
-        Parameters:
-            num_classes: number of object categories, omitting the special no-object category
-            matcher: module able to compute a matching between targets and proposals
-            loss_names: list containing the names of the losses
-            focal_alpha: alpha in Focal Loss
-        '''
-        super().__init__()
-        self.num_classes = num_classes
-        self.matcher = matcher
-        self.loss_names = loss_names
-        self.focal_alpha = focal_alpha  # TODO check
-
-    def forward(self, outputs, targets):
-        ''' This performs the loss computation.
-        Parameters:
-             outputs: dict of tensors, see the output specification of the model for the format
-             targets: list of dicts, such that len(targets) == batch_size.
-                      The expected keys in each dict depends on the losses applied, see each loss' doc
-        '''
-        outputs_without_aux = {k: v for k, v in outputs.items() if k != 'aux_outputs' and k != 'enc_outputs'}
-
-        # Retrieve the matching between the outputs of the last layer and the targets
-        indices = self.matcher(outputs_without_aux, targets)
-
-        # Compute the average number of target boxes accross all nodes, for normalization purposes
-        num_boxes = sum(len(t["labels"]) for t in targets)
-        num_boxes = torch.as_tensor([num_boxes], dtype=torch.float, device=next(iter(outputs.values())).device)
-        if is_dist_avail_and_initialized():
-            torch.distributed.all_reduce(num_boxes)
-        num_boxes = torch.clamp(num_boxes / get_world_size(), min=1).item()
-
-        # Compute all the requested losses
-        losses = self.get_losses(self.loss_names, outputs, targets, indices, num_boxes)
-
-        # In case of auxiliary losses, we repeat this process with the output of each intermediate layer.
-        if 'aux_outputs' in outputs:
-            for i, aux_outputs in enumerate(outputs['aux_outputs']):
-                indices = self.matcher(aux_outputs, targets)
-                loss_names = [k for k in self.loss_names if k != 'masks']  # mask loss is too costly for intermediate layers
-                kwargs = {'log': False} if 'accuracy' in loss_names else {}
-                aux_losses = self.get_losses(loss_names, aux_outputs, targets, indices, num_boxes, **kwargs)
-                aux_losses = {k + f'_{i}': v for k, v in aux_losses.items()}
-                losses.update(aux_losses)
-
-        if 'enc_outputs' in outputs:
-            enc_outputs = outputs['enc_outputs']
-            bin_targets = copy.deepcopy(targets)
-            for bt in bin_targets:
-                bt['labels'] = torch.zeros_like(bt['labels'])
-            indices = self.matcher(enc_outputs, bin_targets)
-            loss_names = [k for k in self.loss_names if k != 'masks']  # mask loss is too costly for intermediate layers
-            kwargs = {'log': False} if 'accuracy' in loss_names else {}
-            enc_losses = self.get_losses(loss_names, enc_outputs, bin_targets, indices, num_boxes, **kwargs)
-            enc_losses = {k + f'_enc': v for k, v in enc_losses.items()}
-            losses.update(enc_losses)
-
-        return losses
-
-    def get_losses(self, loss_names: List[str], outputs, targets, indices, num_boxes, **kwargs):
-        loss_map = {
-            'cls_loss': self.loss_labels,
-            'accuracy': self.loss_labels,
-            'bbox_loss': self.loss_boxes,
-            'giou_loss': self.loss_boxes,
-            'mask_loss': self.loss_masks,
-            'dice_loss': self.loss_masks,
-            'cardinality': self.loss_cardinality,
-        }
-        losses = {}
-        for loss_name in loss_names:
-            if loss_name not in losses:
-                loss_fn = loss_map[loss_name]
-                losses.update(loss_fn(outputs, targets, indices, num_boxes, **kwargs))
-
-        losses = {k: v for k, v in losses.items() if k in losses}
-        return losses
-
-    def loss_labels(self, outputs, targets, indices, num_boxes, log=True):
-        ''' 
-        Classification loss (NLL)
-        targets dicts must contain the key "labels" containing a tensor of dim [nb_target_boxes]
-        '''
-        assert 'pred_logits' in outputs
-        src_logits = outputs['pred_logits']
-
-        idx = self._get_src_permutation_idx(indices)
-        target_classes_o = torch.cat([t["labels"][J] for t, (_, J) in zip(targets, indices)])
-        target_classes = torch.full(src_logits.shape[:2], self.num_classes,
-                                    dtype=torch.int64, device=src_logits.device)
-        target_classes[idx] = target_classes_o
-
-        target_classes_onehot = torch.zeros([src_logits.shape[0], src_logits.shape[1], src_logits.shape[2] + 1],
-                                            dtype=src_logits.dtype, layout=src_logits.layout, device=src_logits.device)
-        target_classes_onehot.scatter_(2, target_classes.unsqueeze(-1), 1)
-
-        target_classes_onehot = target_classes_onehot[:,:,:-1]
-        loss_ce = sigmoid_focal_loss(src_logits, target_classes_onehot, num_boxes, alpha=self.focal_alpha, gamma=2) * src_logits.shape[1]
-        losses = {'cls_loss': loss_ce}
-
-        if log:
-            # TODO this should probably be a separate loss, not hacked in this one here
-            losses['accuracy'] = accuracy(src_logits[idx], target_classes_o)[0]
-        return losses
-
-    @torch.no_grad()
-    def loss_cardinality(self, outputs, targets, indices, num_boxes, log=True):
-        ''' 
-        Compute the cardinality error, ie the absolute error in the number of predicted non-empty boxes
-        This is not really a loss, it is intended for logging purposes only. It doesn't propagate gradients
-        '''
-        pred_logits = outputs['pred_logits']
-        device = pred_logits.device
-        tgt_lengths = torch.as_tensor([len(v["labels"]) for v in targets], device=device)
-        # Count the number of predictions that are NOT "no-object" (which is the last class)
-        card_pred = (pred_logits.argmax(-1) != pred_logits.shape[-1] - 1).sum(1)
-        card_err = F.l1_loss(card_pred.float(), tgt_lengths.float())
-        losses = {'cardinality': card_err}
-        return losses
-
-    def loss_boxes(self, outputs, targets, indices, num_boxes, log=True):
-        ''' 
-        Compute the losses related to the bounding boxes, the L1 regression loss and the GIoU loss
-        targets dicts must contain the key "boxes" containing a tensor of dim [nb_target_boxes, 4]
-        The target boxes are expected in format (center_x, center_y, h, w), normalized by the image size.
-        '''
-        assert 'pred_boxes' in outputs
-        idx = self._get_src_permutation_idx(indices)
-        src_boxes = outputs['pred_boxes'][idx]
-        target_boxes = torch.cat([t['boxes'][i] for t, (_, i) in zip(targets, indices)], dim=0)
-
-        loss_bbox = F.l1_loss(src_boxes, target_boxes, reduction='none')
-
-        losses = {}
-        losses['bbox_loss'] = loss_bbox.sum() / num_boxes
-
-        loss_giou = 1 - torch.diag(box_ops.generalized_box_iou(
-            box_ops.box_cxcywh_to_xyxy(src_boxes),
-            box_ops.box_cxcywh_to_xyxy(target_boxes)))
-        losses['giou_loss'] = loss_giou.sum() / num_boxes
-        return losses
-
-    def loss_masks(self, outputs, targets, indices, num_boxes, log=True):
-        '''
-        Compute the losses related to the masks: the focal loss and the dice loss.
-        targets dicts must contain the key "masks" containing a tensor of dim [nb_target_boxes, h, w]
-        '''
-        assert "pred_masks" in outputs
-
-        src_idx = self._get_src_permutation_idx(indices)
-        tgt_idx = self._get_tgt_permutation_idx(indices)
-
-        src_masks = outputs["pred_masks"]
-
-        # TODO use valid to mask invalid areas due to padding in loss
-        target_masks, valid = nested_tensor_from_tensor_list([t["masks"] for t in targets]).decompose()
-        target_masks = target_masks.to(src_masks)
-
-        src_masks = src_masks[src_idx]
-        # upsample predictions to the target size
-        src_masks = interpolate(src_masks[:, None], size=target_masks.shape[-2:],
-                                mode="bilinear", align_corners=False)
-        src_masks = src_masks[:, 0].flatten(1)
-
-        target_masks = target_masks[tgt_idx].flatten(1)
-
-        losses = {"mask_loss": sigmoid_focal_loss(src_masks, target_masks, num_boxes),
-                  "dice_loss": dice_loss(src_masks, target_masks, num_boxes),
-                  }
-        return losses
-
-    def _get_src_permutation_idx(self, indices):
-        # permute predictions following indices
-        batch_idx = torch.cat([torch.full_like(src, i) for i, (src, _) in enumerate(indices)])
-        src_idx = torch.cat([src for (src, _) in indices])
-        return batch_idx, src_idx
-
-    def _get_tgt_permutation_idx(self, indices):
-        # permute targets following indices
-        batch_idx = torch.cat([torch.full_like(tgt, i) for i, (_, tgt) in enumerate(indices)])
-        tgt_idx = torch.cat([tgt for (_, tgt) in indices])
-        return batch_idx, tgt_idx
-
-
-class PostProcess(nn.Module):
-    """ This module converts the model's output into the format expected by the coco api"""
-
-    @torch.no_grad()
-    def forward(self, outputs, target_sizes):
-        ''' Perform the computation
-        Parameters:
-            outputs: raw outputs of the model
-            target_sizes: tensor of dimension [batch_size x 2] containing the size of each images of the batch
-                          For evaluation, this must be the original image size (before any data augmentation)
-                          For visualization, this should be the image size after data augment, but before padding
-        '''
-        out_logits, out_bbox = outputs['pred_logits'], outputs['pred_boxes']
-
-        assert len(out_logits) == len(target_sizes)
-        assert target_sizes.shape[1] == 2
-
-        prob = out_logits.sigmoid()
-        topk_values, topk_indexes = torch.topk(prob.view(out_logits.shape[0], -1), 100, dim=1)
-        scores = topk_values
-        topk_boxes = topk_indexes // out_logits.shape[2]
-        labels = topk_indexes % out_logits.shape[2]
-        boxes = box_ops.box_cxcywh_to_xyxy(out_bbox)
-        boxes = torch.gather(boxes, 1, topk_boxes.unsqueeze(-1).repeat(1,1,4))
-
-        # and from relative [0, 1] to absolute [0, height] coordinates
-        img_h, img_w = target_sizes.unbind(1)
-        scale_fct = torch.stack([img_w, img_h, img_w, img_h], dim=1)
-        boxes = boxes * scale_fct[:, None, :]
-
-        results = [{'scores': s, 'labels': l, 'boxes': b} for s, l, b in zip(scores, labels, boxes)]
-
-        return results
 
 
 def build_deformable_detr(cfg):
